Break up Property flags to correct alignment
The game lays out Property structures as if they were aligned to
4 bytes. Because we defined property flags to be an 8 byte integer,
we tried to talk to the game using a Property structure that had a
stricter 8 byte alignment rather than the 4-byte alignment the game
uses.

The real word effect was that ByteProperty had its enumeration
field offset by 4 bytes, presumably because the compiler inserted
a total of 4 bytes padding to meet the (incorrect) 8 byte alignment
requirement for Property.

Since align(structure) = max(align(fields)), the fix is to break up
the 8 byte integer field into two 4 byte fields, causing the structure
to have the correct 4 byte alignment.