Ignore the two enums that have duplicate names
ECompareObjectOutputLinkIds doesn't show up in any struct
or class, so ignoring it is fine.

One EFlightMode shows up in UWillowNavigationHandle, and the
other shows up in UBehavior_AISetFlight and UAction_SetFlight.
I need to figure out what I want to place in those fields instead.
Probably just a u8 with a comment explaining where to find the
variants.