Remove unnecessary BufWriter
https://doc.rust-lang.org/stable/std/io/struct.BufWriter.html:
"BufWriter<W> can improve the speed of programs that make small and repeated write calls to the same file or network socket. It does not help when writing very large amounts at once, or writing just one or a few times."

Here, we are writing once the entire SDK scope as a string.