Emit mutable references instead of immutable ones
Back in a similiar hook, I sometimes had to traverse a structure's
nested references to modify a field. I wasn't able to do so if any
of the nested references were immutable, so I had to resort to
hacks.

I'm probably going to run into borrowck headaches down the road
with these nested mutable references.