Silence lint about shadowed variant names
If you declare 'enum MyEnum { A, B }', and have a variable `A: MyEnum`,
then the only way to match on the variants is to fully-qualify the enum.

I'm silencing this lint because we generate methods whose parameter names can
be one of the variant names. We don't match variants within the
generated methods, so this lint is not applicable to us.